requirements.txt
Django==1.5.1
#FeinCMS==1.7.4
Pillow==2.0.0
South==0.7.6
argparse==1.2.1
boto==2.8.0
distribute==0.6.24
dj-database-url==0.2.1
django-fhadmin==1.2.0
#django-mptt==0.5.5
django-storages==1.1.8
feincms-oembed==1.1.0
gunicorn==0.17.2
psycopg2==2.5
python-dateutil==2.1
raven==3.3.2
six==1.3.0
wsgiref==0.1.2